feat: context-usage ring next to the model selector
Surface live context-window usage where it is actionable instead of burying
it in the top-bar More menu, following Claude Code desktop's usage-ring
pattern.

- New ContextUsageRing: text-free arc next to the model selector, single
  accent color normally, red at >=90% usage; hover shows the native HTML
  tooltip "percent / window"; click opens the top-bar session-stats panel.
- Extract the inline session-stats panel from AppShell into a reusable
  SessionStatsPanel (shared by the More menu and the ring's click-through),
  with self-contained copy feedback state.
- Wire contextUsage and onSessionStatsPanelOpen through ChatWindow and
  ChatInput; align percent precision to one decimal everywhere.
- i18n: chat.ctxUsage (en/zh-CN).
