+++ lib/i18n/messages/en.ts
@@ -230,6 +230,7 @@ export const enLocale: LocalePlugin = {
     "chat.compactContext": "Compact context",
     "chat.compacting": "Compacting…",
     "chat.compact": "Compact",
+    "chat.ctxUsage": "Context usage",
     "chat.stopAgent": "Stop agent",
     "chat.stop": "Stop",
     "chat.disableSound": "Disable completion sound",
